# Size optimizations for src/index.mjs