--- a/src/build.mjs
+++ b/src/build.mjs
@@ -2,9 +2,9 @@ const stringify = JSON.stringify;
 
 const MODE_SKIP = 0;
 const MODE_TEXT = 1;
-const MODE_WHITESPACE = 3;
-const MODE_TAGNAME = 4;
-const MODE_ATTRIBUTE = 5;
+const MODE_WHITESPACE = 2;
+const MODE_TAGNAME = 3;
+const MODE_ATTRIBUTE = 4;
 
 /** Create a template function given strings from a tagged template. */
 export const build = (statics) => {
@@ -19,40 +19,34 @@ export const build = (statics) => {
 	let spreadClose = '';
 	let quote = '';
 	let depth = 0;
-	let fallbackPropValue = true;
-	let char, propName, idx;
-
+	let char, propName;
+	
 	const commit = field => {
-		if (mode === MODE_TEXT) {
-			if (field || (buffer = buffer.replace(/^\s*\n\s*|\s*\n\s*$/g,''))) {
-				out += childClose + (field || stringify(buffer));
-				childClose = ',';
-			}
+		if (mode === MODE_TEXT && (field || (buffer = buffer.replace(/^\s*\n\s*|\s*\n\s*$/g,'')))) {
+			out += childClose + (field || stringify(buffer));
+			childClose = ',';
 		}
-		else if (mode === MODE_TAGNAME) {
+		else if (mode === MODE_TAGNAME && (field || buffer)) {
 			out += childClose + 'h(' + (field || stringify(buffer));
 			childClose = ',';
 			mode = MODE_WHITESPACE;
 		}
-		else if (mode === MODE_WHITESPACE && buffer === '...') {
+		else if (mode === MODE_WHITESPACE && buffer === '...' && field) {
 			if (!spreadClose) {
 				props = 'Object.assign(' + (props || '{}');
 			}
-			props += propsClose + ',' + (field || '');
+			props += propsClose + ',' + field;
 			propsClose = '';
 			spreadClose = ')';
 		}
-		else if (mode === MODE_ATTRIBUTE || mode === MODE_WHITESPACE) {
-			if (mode === MODE_WHITESPACE) {
-				propName = buffer;
-				buffer = '';
-			}
-			if (propName) {
-				props += (props ? ',' + (propsClose ? '' : '{') : '{') + stringify(propName) + ':' + (field || stringify(buffer || fallbackPropValue));
-				propsClose = '}';
-				propName = '';
-			}
-			fallbackPropValue = true;
+		else if (mode === MODE_WHITESPACE && buffer && !field) {
+			props += (props ? ',' + (propsClose ? '' : '{') : '{') + stringify(buffer) + ':true';
+			propsClose = '}';
+		}
+		else if (mode === MODE_ATTRIBUTE && propName) {
+			props += (props ? ',' + (propsClose ? '' : '{') : '{') + stringify(propName) + ':' + (field || stringify(buffer));
+			propsClose = '}';
+			propName = '';
 		}
 		buffer = '';
 	};
@@ -68,55 +62,62 @@ export const build = (statics) => {
 		for (let j=0; j<statics[i].length; j++) {
 			char = statics[i][j];
 
-			if (mode === MODE_TEXT && char === '<') {
-				// commit buffer
-				commit();
-				tagClose = spreadClose = propsClose = props = '';
-				if (!(depth++) && out && !outClose) {
-					out = '[' + out;
-					outClose = ']';
-				}
-				mode = MODE_TAGNAME;
-			}
-			else if (mode !== MODE_TEXT && (char === quote || !quote) && (idx = '\'">=/\t\n\r '.indexOf(char)) >= 0) {
-				if (idx < 2) {
-					// char === '"' || char === "'"
-					quote = quote ? '' : char;
-				}
-				else if (idx === 2) {
-					// char === '>'
+			if (mode === MODE_TEXT) {
+				if (char === '<') {
+					// commit buffer
 					commit();
-					if (mode !== MODE_SKIP) {
-						out += ',' + (props + propsClose + spreadClose || null);
+					tagClose = spreadClose = propsClose = props = '';
+					if (!(depth++) && out && !outClose) {
+						out = '[' + out;
+						outClose = ']';
 					}
-					out += tagClose;
-					mode = MODE_TEXT;
-					commit();
-				}
-				else if (idx === 3) {
-					// char === '='
-					mode = MODE_ATTRIBUTE;
-					propName = buffer;
-					buffer = fallbackPropValue = '';
-				}
-				else if (idx === 4) {
-					// char === '/'
-					if (!tagClose) {
-						tagClose = ')';
-						// </foo>
-						if (mode === MODE_TAGNAME && !buffer.trim()) {
-							buffer = '';
-							mode = MODE_SKIP;
-						}
-						depth--;
-					}
+					mode = MODE_TAGNAME;
 				}
 				else {
-					// char is a whitespace
-					// <a disabled>
-					commit();
-					mode = MODE_WHITESPACE;
+					buffer += char;
 				}
+			}
+			else if (quote) {
+				if (char === quote) {
+					quote = '';
+				}
+				else {
+					buffer += char;
+				}
+			}
+			else if (char === '"' || char === "'") {
+				quote = char;
+			}
+			else if (char === '>') {
+				commit();
+				if (mode) {
+					out += ',' + (props + propsClose + spreadClose || null);
+				}
+				out += tagClose;
+				mode = MODE_TEXT;
+			}
+			else if (tagClose) {
+				// Skip the rest of the tag
+			}
+			else if (char === '=') {
+				mode = MODE_ATTRIBUTE;
+				propName = buffer;
+				buffer = '';
+			}
+			else if (char === '/') {
+				// </foo>
+				commit();
+				if (mode === MODE_TAGNAME) {
+					mode = MODE_SKIP;
+					buffer = '';
+				}
+				depth--;
+				tagClose = ')';
+			}
+			else if (char === ' ' || char === '\t' || char === '\n' || char === '\r') {
+				// <a disabled>
+				commit();
+				mode = MODE_WHITESPACE;
 			}
 			else {
 				buffer += char;