This pull request optimizes the size of the minified & compressed library. Mostly it's using the trick is collapsing a state flag and a appendable string together, like `propsClose` previously.

 * Remove `inTag` by using the `mode` variable for it instead (we're not in a tag if `mode === MODE_TEXT`).
 * Spread code repeated `Object.assign`.
 * Keep track if child and tag states with `childClose` and `tagClose`.
 * Replace `propHasValue` with `fallbackPropValue` that can be stringified instead when `buffer` is empty.

These changes together have the following effect on the library size:

**Before change:**
```
Build "htm" to dist:
        649 B: htm.mjs.gz
        596 B: htm.mjs.br
        714 B: htm.umd.js.gz
        645 B: htm.umd.js.br
```

**After change:**
```
Build "htm" to dist:
        605 B: htm.mjs.gz
        546 B: htm.mjs.br
        670 B: htm.umd.js.gz
        599 B: htm.umd.js.br
```