@jviide 2018-12-21
Technically this allows replacing the "under 700 bytes" claim with "under 600 bytes" 😉