@jviide 2018-12-22
The situation after the changes suggested by @mikesherov 🙌

**Before this PR**
```
Build "htm" to dist:
        649 B: htm.mjs.gz
        596 B: htm.mjs.br
        714 B: htm.umd.js.gz
        645 B: htm.umd.js.br
```

**After this PR**
```
Build "htm" to dist:
        600 B: htm.mjs.gz
        537 B: htm.mjs.br
        667 B: htm.umd.js.gz
        592 B: htm.umd.js.br
```