@mikesherov 2018-12-23
Caught one more thing that isn’t in the diff. Can replace the two instances of `quote === 0` with `!quote`