@jviide 2018-12-23
Yup, I tried that as well as replacing `mode === MODE_TEXT` with just `!mode` as I chose `MODE_TEXT = 0` just for that. However, the compressed size actually goes up in some cases and down in some. Here are the numbers after replacing `quote === 0` with `!quote`:

```
Build "htm" to dist:
        600 B: htm.mjs.gz
        543 B: htm.mjs.br
        666 B: htm.umd.js.gz
        593 B: htm.umd.js.br
```