@developit 2018-12-24
I've generally had good luck sticking with stricter equality for performance, and then just using it really consistently to ensure gzip compresses things effectively.