@jviide 2018-12-27
Backported size optimizations from PR #53, the numbers are now the following (benchmarks executed without Jest):

**Before change**
```
Build "htm" to dist:
        600 B: htm.mjs.gz
        537 B: htm.mjs.br
        667 B: htm.umd.js.gz
        592 B: htm.umd.js.br
Creation: 11,887/s, average: 84µs
Usage: 551,669/s, average: 1µs
```

**After change**
```
Build "htm" to dist:
        563 B: htm.mjs.gz
        516 B: htm.mjs.br
        629 B: htm.umd.js.gz
        561 B: htm.umd.js.br
Creation: 12,066/s, average: 82µs
Usage: 574,089/s, average: 1µs
```