@jviide 2018-12-27
Added multi-root support, benchmarks are not affected and the file sizes are:
```
Build "htm" to dist:
        589 B: htm.mjs.gz
        536 B: htm.mjs.br
        657 B: htm.umd.js.gz
        585 B: htm.umd.js.br
```