@jviide 2019-01-16
After the latest changes the code sizes are:
```
Build "htm" to dist:
        561 B: htm.mjs.gz
        518 B: htm.mjs.br
        630 B: htm.umd.js.gz
        566 B: htm.umd.js.br
```

Benchmarks seem to be unaffected.