@jviide 2019-02-20
Closing this in favor of #62. Your size optimizations live on there, @mikesherov! 🙂